MNT: rearrange graph helpers
